Move Invariant trait into kani_core; add AssumeSafe model
Move the definition of the `kani::Invariant` trait and its implementations
for primitive types from the `kani` library into a new `generate_invariant!`
macro in `kani_core`, so that the trait is also available in `no_core` mode
(e.g. for `kani verify-std`). The trait is unchanged and remains available
as `kani::Invariant`.

Additionally, introduce an `assume_safe` model function (registered as
`KaniModel::AssumeSafe`), which assumes that a value respects its type's
safety invariant. The compiler will use this model to constrain the
nondeterministic values generated for automatic harnesses to values that
respect the type's safety invariant.
