+++ library/kani/src/lib.rs
@@ -31,7 +31,6 @@ pub mod bounded_arbitrary;
 #[cfg(feature = "concrete_playback")]
 mod concrete_playback;
 pub mod futures;
-pub mod invariant;
 pub mod iter;
 pub mod shadow;
 pub mod vec;
@@ -40,7 +39,6 @@ mod models;
 
 #[cfg(feature = "concrete_playback")]
 pub use concrete_playback::concrete_playback_run;
-pub use invariant::Invariant;
 
 #[cfg(not(feature = "concrete_playback"))]
 /// NOP `concrete_playback` for type checking during verification mode.
